quest/chapter_03/django_pjt/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2024.10
dateparser==1.2.0
decorator==5.1.1
distro==1.9.0
Django==5.1.5
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
executing==2.1.0
Faker==33.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
matplotlib-inline==0.1.7
moment==0.12.1
mypy-extensions==1.0.0
openai==1.61.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2==2.9.10
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
times==0.7
tomli==2.2.1
toposort==1.10
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
wcwidth==0.2.13
